Fix: 프로필 카드 버그 수정
